Moving some env vars around for Ubuntu's sake
also wrapping the top alias in a "if darwin" statement, since -o doesn't
work on linux (and that order is the default anyway).

I'll admit that the env vars may well be wrong, but it was a quick "get
it working" move that I need to revisit and possibly fix here shortly.